- corrected initialisation opf particles from file with zero velocity - improved closing MPI on error - test option: minimum wait time between MPI calls in optim_ppso_robust.R

+++ R/optim_ppso_robust.R
@@ -2,7 +2,7 @@ optim_ppso_robust <-
 function (objective_function=sample_function, number_of_parameters=2, number_of_particles=40,max_number_of_iterations=5, w=1,  C1=2, C2=2, abstol=-Inf,  reltol=-Inf,  max_wait_iterations=50,
    wait_complete_iteration=FALSE,parameter_bounds=cbind(rep(-1,number_of_parameters),rep(1,number_of_parameters)), Vmax=(parameter_bounds[,2]-parameter_bounds[,1])/3,lhc_init=FALSE,
   #runtime & display parameters
-do_plot=NULL, wait_for_keystroke=FALSE, logfile="ppso.log",projectfile="ppso.pro", save_interval=ceiling(number_of_particles/4),load_projectfile="try",break_file=NULL, nslaves=3          )
+do_plot=NULL, wait_for_keystroke=FALSE, logfile="ppso.log",projectfile="ppso.pro", save_interval=ceiling(number_of_particles/4),load_projectfile="try",break_file=NULL, nslaves=3, waittime=0   )
 # do particle swarm optimization
 {
   
@@ -36,7 +36,7 @@ do_plot=NULL, wait_for_keystroke=FALSE, logfile="ppso.log",projectfile="ppso.pro
 
 eval(parse(text=paste(c("update_tasklist_pso=",deparse(update_tasklist_pso_i)))))  #this creates local version of the function update_tasklist_pso (see explanation there)
 eval(parse(text=paste(c("init_particles=",     deparse(init_particles_i)))))  #this creates local version of the function init_particles (see explanation there)
-if ((!is.null(break_file)) & (file.exists(break_file)))      #delete break_file, if existent
+if ((!is.null(break_file)) && (file.exists(break_file)))      #delete break_file, if existent
   unlink(break_file)   
 
 evals_since_lastsave=0                    #for counting function evaluations since last save of project file
@@ -52,18 +52,20 @@ prepare_mpi_cluster=function(nslaves)
 
   mpi.spawn.Rslaves(nslaves=nslaves)
 
-  .Last <- function(){
-    if (is.loaded("mpi_initialize")){
-        if (mpi.comm.size(1) > 0){
-            print("Please use mpi.close.Rslaves() to close slaves.")
-            mpi.close.Rslaves()
-        }
-        print("Please use mpi.quit() to quit R")
-        .Call("mpi_finalize")
+     .Last <- function(){
+      if (is.loaded("mpi_initialize")){
+          if (mpi.comm.size(1) > 0){
+              #print("Please use mpi.close.Rslaves() to close slaves.")
+              mpi.close.Rslaves()
+          }
+          #print("Please use mpi.quit() to quit R")
+          #.Call("mpi_finalize")
+      }
     }
-  }
+  
+   print(paste(mpi.comm.size()-1,"slaves spawned"))
+   options(error=.Last)     #close rmpi on errors
 
-  print(paste(mpi.comm.size()-1,"slaves spawned"))
   
   #options(error=.Last)     #close rmpi on errors
 
@@ -180,6 +182,12 @@ while ((closed_slaves < nslaves) )
             #print(paste("command sent to slave",slave_id))
           }   
           update_tasklist_pso()   #update particle speeds and positions based on available results
+          
+          starttime=Sys.time()
+          while (as.numeric(Sys.time()-starttime)<waittime)
+          {
+          }
+          
           tobecomputed=status==0
          flush.console()
       }
